requirements.txt
Flask==2.3.2
neo4j==5.12.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0